Changed inputs to floats rather than ints

+++ onePlusOne.py
@@ -1,7 +1,7 @@
-e = 1
-f = 1
+e = 1.0
+f = 1.0
 
-def onePlusOne(e: int, f: int):
+def onePlusOne(e: float, f: float):
     return e + f
 
 print(onePlusOne(e, f))
